optimize

+++ src/routes/rancangan.ts
@@ -2,7 +2,7 @@ import { genericResponse } from "@/constants.ts";
 import { server } from "@/index.ts";
 import { permintaanTraining } from "@/models/draft_permintaan_training.ts";
 import { pelaksanaanPelatihan, pelaksanaanPelatihanSchema } from "@/models/rancangan_pelatihan.ts";
-import { ruangan, statusRuangan } from "@/models/ruangan.ts";
+import { ruangan } from "@/models/ruangan.ts";
 import { tablePeserta } from "@/models/table_peserta.ts";
 import { db } from "@/modules/database.ts";
 import { eq } from "drizzle-orm";
@@ -27,7 +27,7 @@ export const route = (instance: typeof server) => {
                     401: genericResponse(401)
                 }
             }
-        }, async (req) => {
+        }, async () => {
             const res = await db.select().from(pelaksanaanPelatihan).execute();
             return {
                 statusCode: 200,
@@ -50,26 +50,26 @@ export const route = (instance: typeof server) => {
                 }
             }
         }, async (req) => {
-            const { id_pelatihan, id_instruktur, id_ruangan, tanggal_mulai, tanggal_selesai, jam_mulai, jam_selesai, jenis_training} = req.body;
-            const result = await db.insert(pelaksanaanPelatihan).values({ 
-                id_pelatihan, 
-                id_instruktur, 
+            const { id_pelatihan, id_instruktur, id_ruangan, tanggal_mulai, tanggal_selesai, jam_mulai, jam_selesai, jenis_training } = req.body;
+            const result = await db.insert(pelaksanaanPelatihan).values({
+                id_pelatihan,
+                id_instruktur,
                 tanggal_mulai,
-                tanggal_selesai, 
-                jam_mulai, 
-                jam_selesai, 
-                jenis_training, 
-                id_ruangan, 
-                is_selesai: "belum", 
-                createdAt: new Date() 
+                tanggal_selesai,
+                jam_mulai,
+                jam_selesai,
+                jenis_training,
+                id_ruangan,
+                is_selesai: "belum",
+                createdAt: new Date()
             }).returning().execute();
 
             const id_pelaksanaanPelatihan = result[0].id;
 
             await db.insert(permintaanTraining).values({
                 id_pelaksanaanPelatihan,
-                status: "menunggu",
-            })
+                status: "menunggu"
+            });
 
             await db.update(ruangan).set({
                 status_ruangan: "dipakai"
@@ -95,13 +95,13 @@ export const route = (instance: typeof server) => {
                 }
             }
         }, async (req) => {
-            const { id, id_instruktur, id_ruangan, tanggal_mulai, tanggal_selesai} = req.body;
+            const { id, id_instruktur, id_ruangan, tanggal_mulai, tanggal_selesai } = req.body;
 
             await db.update(pelaksanaanPelatihan).set({
                 id_instruktur,
                 tanggal_mulai,
                 tanggal_selesai,
-                id_ruangan,
+                id_ruangan
             }).where(eq(pelaksanaanPelatihan.id, Number(id))).execute();
             return {
                 statusCode: 200,
@@ -123,10 +123,26 @@ export const route = (instance: typeof server) => {
                 }
             }
         }, async (req) => {
-            const {id} = req.body;
-            await db.delete(permintaanTraining) .where(eq(permintaanTraining.id_pelaksanaanPelatihan, Number(id))) .execute();
-            await db.delete(tablePeserta) .where(eq(tablePeserta.id_pelaksanaan_pelatihan, Number(id))) .execute();
-            await db.delete(pelaksanaanPelatihan) .where(eq(pelaksanaanPelatihan.id, Number(id))) .execute();
+            const { id } = req.body;
+
+            const training = await db.select().from(pelaksanaanPelatihan).where(eq(pelaksanaanPelatihan.id, Number(id))).execute();
+
+            if (!training) {
+                return {
+                    statusCode: 401,
+                    message: "data not found"
+                };
+            }
+
+            await db.delete(ruangan).where(eq(ruangan.id, training[0].id_ruangan)).execute();
+            await db.delete(permintaanTraining).where(eq(permintaanTraining.id_pelaksanaanPelatihan, Number(id))).execute();
+            await db.delete(tablePeserta).where(eq(tablePeserta.id_pelaksanaan_pelatihan, Number(id))).execute();
+            await db.delete(pelaksanaanPelatihan).where(eq(pelaksanaanPelatihan.id, Number(id))).execute();
+
+            await db.update(ruangan).set({
+                status_ruangan: "tidak dipakai"
+            }).where(eq(ruangan.id, Number(id))).execute();
+
             return {
                 statusCode: 200,
                 message: "Success"
@@ -141,7 +157,7 @@ export const route = (instance: typeof server) => {
                     authorization: z.string().transform(v => v.replace("Bearer ", ""))
                 }),
                 body: z.object({
-                    id: z.string(),
+                    id: z.string()
                 }),
                 response: {
                     200: genericResponse(200),
@@ -149,10 +165,25 @@ export const route = (instance: typeof server) => {
                 }
             }
         }, async (req) => {
-            const {id} = req.body;
+            const { id } = req.body;
+
+            const training = await db.select().from(pelaksanaanPelatihan).where(eq(pelaksanaanPelatihan.id, Number(id))).execute();
+
+            if (!training) {
+                return {
+                    statusCode: 401,
+                    message: "data not found"
+                };
+            }
+
+            await db.update(ruangan).set({
+                status_ruangan: "tidak dipakai"
+            }).where(eq(ruangan.id, training[0].id_ruangan)).execute();
+
             await db.update(pelaksanaanPelatihan).set({
-                is_selesai: "selesai",
+                is_selesai: "selesai"
             }).where(eq(pelaksanaanPelatihan.id, Number(id))).execute();
+
             return {
                 statusCode: 200,
                 message: "Success"
@@ -193,5 +224,5 @@ export const route = (instance: typeof server) => {
                 data: trainingDetail[0]
             };
         }
-        )
+        );
 };
